Add files via upload
The Substance Use Prevention and Intervention executives wanted a high level summary of the changes in the number of people served, locations, etc. since changes to the service delivery system occurred in 2015. This one page infographic has been used widely in Riverside County to demonstrate the impact the waivered service system delivery has had, and is also being shared at the state level.