Add files via upload
An evaluation of the Recovery Residence program in substance abuse prevention and treatment. In order to complete this request, I had to write a complex query to get the data from our electronic health record, determine which people qualified as a comparison group, cleaned the data, and then did analysis.